split qwen36 moe decode session setup
