+++ crates/runner/src/main.rs
@@ -23,6 +23,7 @@ mod qwen36_moe_mtp_loader;
 mod qwen36_moe_persistent_decode;
 mod qwen36_moe_prefetch;
 mod qwen36_moe_residency;
+mod qwen36_moe_session;
 mod qwen36_moe_speculative;
 mod qwen36_moe_state;
 mod qwen36_moe_telemetry;
