+++ crates/runner/src/qwen36_moe_engine.rs
@@ -13,24 +13,23 @@
 use std::path::Path;
 
 use anyhow::{anyhow, Context, Result};
-use gpu_hal::{set_backend, Backend, GpuBuffer, ScalarType, VirtualArena};
+use gpu_hal::{set_backend, Backend, VirtualArena};
 use model_store::BakedStore;
 use qwen36_moe::config::TextConfig;
 
 use crate::qwen36_moe_decode::{
     argmax_bf16_logits, host_final_norm_lm_head, run_chained_decode, run_chained_decode_fast,
     run_chained_decode_fast_with_expert_prefetch, sample_bf16_logits, ExpertPrefetchPhase,
-    ExpertRoute, LayerBuffers, MtpLayerBuffers, MultiLayerGeom, XorshiftRng,
+    ExpertRoute, LayerBuffers, MultiLayerGeom, XorshiftRng,
 };
 use crate::qwen36_moe_dry_run::{
     print_report, run_qwen36_moe_dry_run, ContextSizeSource, DryRunReport,
 };
 use crate::qwen36_moe_host::{host_load_bytes, load_lm_head_bf16, lookup_embed_row};
-use crate::qwen36_moe_layers::{
-    load_all_layer_buffers, load_layer_buffers, load_to_gpu, Qwen36WeightMode,
-};
+use crate::qwen36_moe_layers::{load_all_layer_buffers, load_layer_buffers, Qwen36WeightMode};
 use crate::qwen36_moe_prefetch::handle_moe_expert_prefetch;
 use crate::qwen36_moe_residency::{MoeExpertResidencyConfig, MoeExpertResidencyManager};
+use crate::qwen36_moe_session::{prepare_decode_session, Qwen36DecodeSession};
 use crate::qwen36_moe_speculative::{
     run_speculative_decode_step, run_speculative_decode_step_batched,
 };
@@ -247,20 +246,6 @@ struct Qwen36DecodeLayers {
     layers: Vec<LayerBuffers>,
     moe_expert_arena: Option<VirtualArena>,
     moe_expert_residency: Option<MoeExpertResidencyManager>,
-}
-
-struct Qwen36DecodeSession {
-    final_norm_w_buf: GpuBuffer,
-    lm_head_w_buf: GpuBuffer,
-    logits_buf: GpuBuffer,
-    counter_buf: GpuBuffer,
-    final_hidden_buf: GpuBuffer,
-    mtp_buffers: Option<MtpLayerBuffers>,
-    mtp_forward_scratch: Option<crate::qwen36_moe_mtp::MtpForwardScratch>,
-    mtp_chain_scratch: Option<crate::qwen36_moe_mtp::MtpChainScratch>,
-    embed_w_buf: Option<GpuBuffer>,
-    linear_attn_snapshot: Option<crate::qwen36_moe_state::LinearAttnSnapshot>,
-    persistent_scratch: Option<crate::qwen36_moe_persistent_decode::PersistentScratch>,
 }
 
 #[derive(Default)]
@@ -486,163 +471,6 @@ fn load_decode_layers_with_vmm_strategy(
         layers,
         moe_expert_arena: None,
         moe_expert_residency: None,
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn prepare_decode_session(
-    store: &BakedStore,
-    ordinal: usize,
-    geom: &MultiLayerGeom,
-    text_config: &TextConfig,
-    weight_prefix: &str,
-    kv_max_t: usize,
-    speculative_decode: bool,
-    batched_spec_verify: bool,
-    persistent_decode: bool,
-    layers: &mut Vec<LayerBuffers>,
-) -> Result<Qwen36DecodeSession> {
-    let mtp_buffers_opt = if speculative_decode {
-        match crate::qwen36_moe_mtp_loader::load_mtp_buffers(store, ordinal, geom, kv_max_t)
-            .context("load MTP head from bake")?
-        {
-            Some(mtp) => {
-                println!(
-                    "  MTP head: loaded 19 mtp.* tensors (~1.6 GiB BF16) — \
-                     speculative draft + sequential-verify path active. \
-                     NOTE: sequential verification has zero amortized speedup \
-                     over plain greedy decode (Phase 6.4's batched verify \
-                     kernel is what delivers throughput); this path is the \
-                     correctness foundation."
-                );
-                Some(mtp)
-            }
-            None => {
-                anyhow::bail!(
-                    "--speculative-decode requested but the bake doesn't \
-                     include mtp.* tensors. Re-bake against the post-#84 \
-                     `oracle/bake_int4.py`, or pull the new release tarball \
-                     once the producer workflow at GitHub issue #87 lands."
-                );
-            }
-        }
-    } else {
-        None
-    };
-
-    let final_norm_bytes = host_load_bytes(store, &format!("{weight_prefix}.norm.weight"))
-        .context("load final norm")?;
-    let lm_head_bf16_bytes = load_lm_head_bf16(store, text_config, weight_prefix, geom)
-        .context("prepare lm_head BF16 buffer")?;
-    println!(
-        "  uploading lm_head BF16 ({:.1} MiB) and final norm ({:.1} KiB) to GPU…",
-        lm_head_bf16_bytes.len() as f64 / MIB,
-        final_norm_bytes.len() as f64 / 1024.0,
-    );
-    let final_norm_w_buf = GpuBuffer::from_host_bytes(
-        ordinal,
-        ScalarType::BF16,
-        &[geom.hidden as usize],
-        &final_norm_bytes,
-    )
-    .context("upload final_norm_w to GPU")?;
-    let lm_head_w_buf = GpuBuffer::from_host_bytes(
-        ordinal,
-        ScalarType::BF16,
-        &[geom.vocab as usize, geom.hidden as usize],
-        &lm_head_bf16_bytes,
-    )
-    .context("upload lm_head BF16 to GPU")?;
-    let logits_buf = GpuBuffer::zeros(ordinal, ScalarType::BF16, &[geom.vocab as usize])
-        .context("alloc logits_buf on GPU")?;
-    let counter_buf = GpuBuffer::zeros(ordinal, ScalarType::U32, &[1])
-        .context("alloc lm_head counter_buf on GPU")?;
-    let final_hidden_buf = GpuBuffer::zeros(ordinal, ScalarType::BF16, &[geom.hidden as usize])
-        .context("alloc final_hidden_buf on GPU")?;
-    drop(lm_head_bf16_bytes);
-    drop(final_norm_bytes);
-
-    let mtp_buffers = mtp_buffers_opt;
-    let mtp_forward_scratch = if mtp_buffers.is_some() {
-        Some(
-            crate::qwen36_moe_mtp::alloc_mtp_forward_scratch(ordinal, geom, kv_max_t)
-                .context("alloc MTP forward scratch")?,
-        )
-    } else {
-        None
-    };
-    let mtp_chain_scratch = if mtp_buffers.is_some() {
-        Some(
-            crate::qwen36_moe_mtp::alloc_mtp_chain_scratch(ordinal, geom)
-                .context("alloc MTP chain scratch")?,
-        )
-    } else {
-        None
-    };
-    let embed_w_buf = if mtp_buffers.is_some() {
-        let embed_name = format!("{weight_prefix}.embed_tokens.weight");
-        let embed = load_to_gpu(store, ordinal, &embed_name)
-            .with_context(|| format!("upload {embed_name} to GPU"))?;
-        println!(
-            "  uploaded embed_tokens ({:.0} MiB BF16) and allocated MTP \
-             scratches (K={} drafts/step)",
-            (geom.vocab as f64 * geom.hidden as f64 * 2.0) / MIB,
-            QWEN36_NUM_SPECULATIVE_TOKENS,
-        );
-        Some(embed)
-    } else {
-        None
-    };
-
-    let linear_attn_snapshot = if speculative_decode && batched_spec_verify {
-        Some(
-            crate::qwen36_moe_state::save_linear_attn_state(ordinal, layers)
-                .context("alloc linear-attn state snapshot for batched spec verify")?,
-        )
-    } else {
-        None
-    };
-    if linear_attn_snapshot.is_some() {
-        println!(
-            "  --batched-spec-verify: linear-attn state snapshot allocated \
-             (K+1 chains run per spec iter; restore + replay accepted prefix \
-             on partial accept)"
-        );
-    }
-
-    let persistent_scratch = if persistent_decode {
-        let scratch =
-            crate::qwen36_moe_persistent_decode::PersistentScratch::new(ordinal, geom, layers)
-                .context("alloc PersistentScratch for --persistent-decode")?;
-        println!(
-            "  --persistent-decode: megakernel scratch allocated \
-             (descs={}KiB, workspace={}KiB, ping/pong={}KiB){}",
-            scratch.layer_descs_dev.len_bytes() / 1024,
-            scratch.workspace.len_bytes() / 1024,
-            scratch.hidden_ping.len_bytes() / 1024,
-            if speculative_decode {
-                " — also routes spec-verify chains through persistent"
-            } else {
-                ""
-            },
-        );
-        Some(scratch)
-    } else {
-        None
-    };
-
-    Ok(Qwen36DecodeSession {
-        final_norm_w_buf,
-        lm_head_w_buf,
-        logits_buf,
-        counter_buf,
-        final_hidden_buf,
-        mtp_buffers,
-        mtp_forward_scratch,
-        mtp_chain_scratch,
-        embed_w_buf,
-        linear_attn_snapshot,
-        persistent_scratch,
     })
 }
